Add files via upload
This project is a dynamic website for "Wadi Degla Club," highlighting its tennis academy and community initiatives. I built it using HTML and CSS, ensuring a fully responsive design that adapts seamlessly across devices. The site features an engaging layout with sections for academy details, an interactive schedule, and a news gallery, providing an excellent user experience with smooth navigation and mobile-friendly functionality.